third_party/WebKit/LayoutTests/platform/win/custom-elements/constructor-context-dies-cross-context-call-expected.txt

This is a testharness.js-based test.
PASS context is destroyed; call across context 
Harness: the test ran to completion.
